pyproject.toml
[project]
name = "bind-tools"
version = "0.1.0"
description = "Protein-ligand binding analysis toolkit"
requires-python = ">=3.9,<3.13"
dependencies = [
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "rcsbsearchapi>=2.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
bind-api = "bind_tools.api:main"
bind-agent = "bind_tools.agent.cli:main"
bind-boltz = "bind_tools.boltz.cli:app"
bind-gnina = "bind_tools.gnina.cli:app"
bind-ligand = "bind_tools.ligand.cli:main"
bind-ligprep = "bind_tools.ligprep.cli:main"
bind-plip = "bind_tools.plip.cli:app"
bind-posebusters = "bind_tools.posebusters.cli:app"
bind-protein = "bind_tools.protein.cli:main"
bind-protprep = "bind_tools.protprep.cli:main"
bind-qmd = "bind_tools.qmd.cli:app"
bind-resolve = "bind_tools.resolve.cli:app"
bind-search = "bind_tools.search.cli:main"
bind-memory = "bind_tools.memory.cli:app"
bind-websearch = "bind_tools.websearch.cli:main"

[project.optional-dependencies]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "psycopg2-binary>=2.9.0",
]
db = [
    "psycopg2-binary>=2.9.0",
]
modal = [
    "modal>=0.73.0",
    "python-dotenv>=1.0.0",
]
protprep = [
    "pdbfixer>=1.9",
    "openmm>=8.0",
    "pdb2pqr>=3.6.0",
]
batch = [
    "rdkit>=2023.0",
]
ligprep = [
    "rdkit>=2023.0",
    "meeko>=0.5.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "psycopg2-binary>=2.9.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A"]
